handle errors

+++ controllers/Chats.js
@@ -28,8 +28,12 @@ const returnAllChats = async (req, res) => {
         const username = getUserNameFromToken(req.headers.authorization)
         if (username !== "Invalid Token") {
             const allChats = await chatsService.returnAllChats(username);
-            res.status(200).send(allChats);
-            return
+            if (allChats != -1) {
+                res.status(200).send(allChats);
+                return
+            } else {
+                res.status(400).send('failed. problem with the DB');
+            }
         } else {
             return res.status(401).send("Invalid Token");
         }
@@ -40,15 +44,19 @@ const returnAllChats = async (req, res) => {
 }
 
 const createChat = async (req, res) => {
-    const requestBody = req.body; // Assuming the JSON object is in the request body
-    // get the login username
-    const username = getUserNameFromToken(req.headers.authorization)
-    if (username !== "Invalid Token") {
-        const newChat = await chatsService.createChat(requestBody.username, username);
-        if (newChat != -1) {
-            res.status(200).json(newChat);
+    if (req.headers.authorization) {
+        const requestBody = req.body; // Assuming the JSON object is in the request body
+        // get the login username
+        const username = getUserNameFromToken(req.headers.authorization)
+        if (username !== "Invalid Token") {
+            const newChat = await chatsService.createChat(requestBody.username, username);
+            if (newChat != -1) {
+                res.status(200).json(newChat);
+            } else {
+                res.status(400).send('failed. problem with the DB');
+            }
         } else {
-            res.status(400).send('failed. problem with the DB');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -56,14 +64,19 @@ const createChat = async (req, res) => {
 }
 
 const returnTheConversation = async (req, res) => {
-    // get the login username
-    const username = getUserNameFromToken(req.headers.authorization)
-    if (username !== "Invalid Token") {
-        const allMessages = await chatsService.returnTheConversation(req.params.id, username)
-        if (allMessages != -1) {
-            res.status(200).json(allMessages);
+    if (req.headers.authorization) {
+        // get the login username
+        const username = getUserNameFromToken(req.headers.authorization)
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const allMessages = await chatsService.returnTheConversation(id)
+            if (allMessages != -1) {
+                res.status(200).json(allMessages);
+            } else {
+                res.status(400).send('failed. problem with the DB');
+            }
         } else {
-            res.status(400).send('failed. problem with the DB');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -71,14 +84,19 @@ const returnTheConversation = async (req, res) => {
 }
 
 const addNewMessage = async (req, res) => {
-    const content = req.body.msg;
-    const username = getUserNameFromToken(req.headers.authorization);
-    if (username !== "Invalid Token") {
-        const newMessage = await chatsService.addNewMessage(username, content, req.params.id);
-        if (newMessage != -1) {
-            res.status(200).json(newMessage);
+    if (req.headers.authorization) {
+        const content = req.body.msg; //get the content of the message
+        const username = getUserNameFromToken(req.headers.authorization);
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const newMessage = await chatsService.addNewMessage(username, content, id);
+            if (newMessage != -1) {
+                res.status(200).json(newMessage);
+            } else {
+                res.status(400).send('failed. problem with the DB');
+            }
         } else {
-            res.status(400).send('failed. problem with the DB');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -86,13 +104,18 @@ const addNewMessage = async (req, res) => {
 }
 
 const returnAllTheMessages = async (req, res) => {
-    const username = getUserNameFromToken(req.headers.authorization);
-    if (username !== "Invalid Token") {
-        const allTheMessages = await chatsService.returnAllTheMessages(req.params.id);
-        if (allTheMessages != -1) {
-            res.status(200).json(allTheMessages);
+    if (req.headers.authorization) {
+        const username = getUserNameFromToken(req.headers.authorization);
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const allTheMessages = await chatsService.returnAllTheMessages(id);
+            if (allTheMessages != -1) {
+                res.status(200).json(allTheMessages);
+            } else {
+                res.status(400).send('failed. problem with the DB');
+            }
         } else {
-            res.status(400).send('failed. problem with the DB');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -100,18 +123,18 @@ const returnAllTheMessages = async (req, res) => {
 }
 
 const updateChats = async (req, res) => {
-    const username = getUserNameFromToken(req.headers.authorization);
-    const id = req.params.id;
-    if (username !== "Invalid Token") {
-        const chat = await Chats.findOne({ id });
-        const userOne = chat.users[0].username
-        const userTwo = chat.users[1].username
-        const updatedChatsOne = await chatsService.updateChats(userOne ,id);
-        const updatedChatsTwo = await chatsService.updateChats(userTwo ,id);
-        if (updatedChatsOne != -1 && updatedChatsTwo != -1) {
-            res.status(200).send("success");
+    if (req.headers.authorization) {
+        const username = getUserNameFromToken(req.headers.authorization);
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const isUpdated = await chatsService.updateChats(id);
+            if (isUpdated != -1) {
+                res.status(200).send("success");
+            } else {
+                res.status(400).send('failed. problem with the DB');
+            }
         } else {
-            res.status(400).send('failed. problem with the DB');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -119,14 +142,18 @@ const updateChats = async (req, res) => {
 }
 
 const deleteChat = async (req, res) => {
-    const username = getUserNameFromToken(req.headers.authorization);
-    const id = req.params.id;
-    if (username !== "Invalid Token") {
-        const isDeleted = await chatsService.deleteChat(username, id);
-        if (isDeleted != -1) {
-            res.status(200).json(isDeleted)
+    if (req.headers.authorization) {
+        const username = getUserNameFromToken(req.headers.authorization);
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const isDeleted = await chatsService.deleteChat(username, id);
+            if (isDeleted != -1) {
+                res.status(200).json(isDeleted)
+            } else {
+                res.status(404).send('Not Found');
+            }
         } else {
-            res.status(404).send('Not Found');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -134,13 +161,17 @@ const deleteChat = async (req, res) => {
 }
 
 const getNotifications = async (req, res) => {
-    const username = getUserNameFromToken(req.headers.authorization);
-    if (username !== "Invalid Token") {
-        const notifications = await chatsService.getNotifications(username);
-        if (notifications != -1) {
-            res.status(200).send(notifications);
+    if (req.headers.authorization) {
+        const username = getUserNameFromToken(req.headers.authorization);
+        if (username !== "Invalid Token") {
+            const notifications = await chatsService.getNotifications(username);
+            if (notifications != -1) {
+                res.status(200).send(notifications);
+            } else {
+                res.status(404).send('Not Found');
+            }
         } else {
-            res.status(404).send('Not Found');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -148,14 +179,18 @@ const getNotifications = async (req, res) => {
 }
 
 const addNotification = async (req, res) => {
-    const username = getUserNameFromToken(req.headers.authorization);
-    const id = req.params.id;
-    if (username !== "Invalid Token") {
-        const okay = await chatsService.addNotification(username, id);
-        if (okay != -1) {
-            res.status(200).json(okay);
+    if (req.headers.authorization) {
+        const username = getUserNameFromToken(req.headers.authorization);
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const okay = await chatsService.addNotification(username, id);
+            if (okay != -1) {
+                res.status(200).json(okay);
+            } else {
+                res.status(404).send('Could not add notifications');
+            }
         } else {
-            res.status(404).send('Not Found');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
@@ -163,14 +198,18 @@ const addNotification = async (req, res) => {
 }
 
 const resetNotifications = async (req, res) => {
-    const username = getUserNameFromToken(req.headers.authorization);
-    const id = req.params.id;
-    if (username !== "Invalid Token") {
-        const okay = await chatsService.resetNotifications(username, id);
-        if (okay != -1) {
-            res.status(200).json(okay);
+    if (req.headers.authorization) {
+        const username = getUserNameFromToken(req.headers.authorization);
+        const id = req.params.id;
+        if (username !== "Invalid Token") {
+            const okay = await chatsService.resetNotifications(username, id);
+            if (okay != -1) {
+                res.status(200).json(okay);
+            } else {
+                res.status(404).send('Could not reset notifications');
+            }
         } else {
-            res.status(404).send('Not Found');
+            return res.status(401).send("Invalid Token");
         }
     } else {
         return res.status(401).send('Token required');
